+++ docs/release-notes.md
@@ -27,6 +27,9 @@ title: jqwik Release Notes
 
 #### New and Enhanced Features
 
+- More work on 
+  [generation of DateTimes](/docs/snapshot/user-guide.html#generation-of-datetimes)
+
 #### Breaking Changes
 
 #### Bug Fixes